more robust environment auto-switching
This commit removes Ember from vendor/assets/javascripts completely. It
is important to do that because Rails will automatically add that
entire tree to the asset pipeline. That makes it possible to get the
production version of Ember when you were expecting the development
version of Ember and vice versa, depending on exactly how the asset
paths are ordered and whether you sacrificed the correct farmyard
animal to Sprockets within the last month.

In addition, this commit provides a map between Rails environments and the version of Ember that ember-rails will automatically provide. If you are, for example, using capybara-webkit to do acceptance testing, you absolutely want to run your acceptance tests against the production version of Ember. The previous behaviour was to provide the development version of Ember test environments, which is IMO pretty dangerous.