Update README.md
Add minor notice on how the local cache must be cleared when the value of precompile is changed.

The programatical solution to this is to condition the cache on the value of this setting. Right now it's only a file fingerprint, but if this can be changed to a union of the fingerprint and the configuration that would fix it.